Code cleanup
move impl to .cpp
